Context Aware Feature Implemented
For the implementation of the Context Aware feature I used the flutter light plugin and a stream subscription to listen the data calculated by the light sensor. In this way, when the calculated lightLevel was low, a Dark Mode was suggested to the user by an Alert Dialog. If the suggestion was accepted, the application's theme changed automatically. I also used the Shared preferences plugin in order to save in cache either the dark or light mode. I used the Future function to concurrently access the preferences saved in the previous session. All the theme information and the cache management are in the theme.dart file.  the observer pattern was also implemented, because the class MyApp consumes the class ThemeNotifier, so that when the preferences are changed, the class ThemeNotifier uses the metod notifyListeners(), and MyApp reacts to the notification.