UI Enhancement and Connectivity Service
The trending products list was displayed in a vertical container, and now it is in an horizontal one. Overflow problems were also fixed, and now the user can click the entire box of the product. A service was created for managing the connectivity, but it is still missing its implementation.