Eventual Connectivity implemented
For the detection of connectivity I used the connectivity_plus plugin. The listening was made using a Stream linked to the Connectivity Service class. The Search View listens this stream, and if detects that there is not connection, changes the grid displaying the products. It was also implemented a filter bar for the search, so that it was more efficient.