Advances in implementation
The Search View was changed so that when the user clicks on a product, is redirected to the ProductDetail view. The  ProductDetail view was completely improved, the code was ordered and the theme consumption was added. The information of the seller was added, but it is still missing the retrieving from the database.