Publish View enhanced
Asynchronous tasks and connectivity manipulation were implemented. A stream was created so that when there is no internet connection, the user can still publish the product, but it is only reflected in the database when the connection is restored. All the verifications of each field were also implemented, so that the input values are restricted